[MGDOBR-1237] Split shard operator integration tests into v1 and v2 (#1373)
* Split shard operator integration tests into v1 and v2

* Rebase and Split shard operator integration tests into v1 and v2
