+++ .github/workflows/Operator-CI-v1.yaml
@@ -1,11 +1,11 @@
-name: Shard Operator - CI
+name: Shard Operator - CI - v1
 on:
   pull_request_target:
     types: [ opened, labeled, unlabeled, synchronize ]
     paths:
-      - '.github/workflows/Operator-CI.yaml'
+      - '.github/workflows/Operator-CI-v1.yaml'
       - 'shard-operator-parent/**'
-      - 'integration-tests/shard-operator-integration-tests/**'
+      - 'integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/**'
 jobs:
   start-runner:
     name: Start self-hosted EC2 runner
@@ -120,13 +120,13 @@ jobs:
         uses: ./.github/actions/maven
         with:
           maven-command: clean verify -Pcucumber -Dparallel -Dit.shard.kafka.bootstrap.servers=$IT_SHARD_KAFKA_BOOTSTRAP_SERVERS -Dit.shard.kafka.user=$IT_SHARD_KAFKA_USER -Dit.shard.kafka.password=$IT_SHARD_KAFKA_PASSWORD -Dit.shard.kafka.topic.name=$IT_SHARD_KAFKA_TOPIC_NAME -Dit.shard.kafka.protocol=SASL_SSL -Dit.shard.kafka.sasl.mechanism=PLAIN
-          working-directory: integration-tests/shard-operator-integration-tests
+          working-directory: integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1
       - name: Archive test logs
         if: always()
         uses: actions/upload-artifact@v3
         with:
           name: cucumber-logs
-          path: integration-tests/shard-operator-integration-tests/target/logs/**/*.*
+          path: integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/target/logs/**/*.*
       - name: Print operator log
         if: always()
         run: |
