+++ integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/run-local-tests.sh
