+++ integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/src/test/filtered-resources/reportportal.properties
