+++ integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/src/test/java/com/redhat/service/smartevents/shard/operator/cucumber/RunCucumberTest.java
