+++ integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/src/test/java/com/redhat/service/smartevents/shard/operator/cucumber/steps/BridgeIngressSteps.java
@@ -8,11 +8,11 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeoutException;
 
+import com.redhat.service.smartevents.integration.tests.common.TimeUtils;
+import com.redhat.service.smartevents.integration.tests.common.Utils;
+import com.redhat.service.smartevents.integration.tests.context.ShardContext;
 import com.redhat.service.smartevents.shard.operator.core.providers.GlobalConfigurationsConstants;
 import com.redhat.service.smartevents.shard.operator.core.utils.LabelsBuilder;
-import com.redhat.service.smartevents.shard.operator.cucumber.common.Context;
-import com.redhat.service.smartevents.shard.operator.cucumber.common.TimeUtils;
-import com.redhat.service.smartevents.shard.operator.cucumber.utils.Utils;
 import com.redhat.service.smartevents.shard.operator.v1.resources.BridgeIngress;
 
 import io.cucumber.java.en.Then;
@@ -26,9 +26,9 @@ import io.fabric8.kubernetes.api.model.SecretBuilder;
  */
 public class BridgeIngressSteps {
 
-    private Context context;
+    private ShardContext context;
 
-    public BridgeIngressSteps(Context context) {
+    public BridgeIngressSteps(ShardContext context) {
         this.context = context;
     }
