+++ integration-tests/shard-operator-integration-tests/shard-operator-integration-tests-v1/src/test/resources/junit-platform.properties
